refactor(cli): split p2p network record helpers
