+++ crates/gents-cli/src/commands/p2p/mod.rs
@@ -5,8 +5,10 @@ mod connect;
 mod documents;
 mod invite;
 mod join;
+mod membership_records;
 mod network;
 mod network_admin;
+mod network_records;
 mod output;
 mod pairings;
 mod replicators;
